Bump polliwog

+++ lacecore/_selection/selection_mixin.py
@@ -236,7 +236,6 @@ class SelectionMixin:
         See also:
             https://polliwog.readthedocs.io/en/latest/#polliwog.Plane
         """
-        import numpy as np
         from polliwog import Plane
         from polliwog.plane import slice_triangles_by_plane
         from .._mesh import Mesh
@@ -245,9 +244,8 @@ class SelectionMixin:
 
         vertices, faces = slice_triangles_by_plane(
             vertices=self.v,
-            # TODO: Fix this in polliwog.
-            faces=self.f.astype(np.uint64),
+            faces=self.f,
             point_on_plane=plane.reference_point,
             plane_normal=plane.normal,
         )
-        return Mesh(v=vertices, f=faces.astype(np.int64))
+        return Mesh(v=vertices, f=faces)
